fix: 修复 SpaceBattle code review 发现的问题
- 英文注释改为简体中文（用户全局规则）
- 提取重复的 Percentile 到共享工具类
- US47 速度区间改为闭区间 [0,200]
- 删除未使用的 RulesetVersion 字段
- ModeStartedTick uint→long 消除换行隐患
- US110 增强 SpatialBarrierOnly 防护测试
- DESIGN.md 速度/阶段表述修正

+++ demo/SpaceBattle/SpaceBattle.Main/Components.cs
@@ -59,7 +59,7 @@ public struct Behavior
     [Field] public byte Mode;
     [Field] public byte Phase;
     [Field] public ushort TicksRemaining;
-    [Field] public uint ModeStartedTick;
+    [Field] public long ModeStartedTick;
 }
 
 public enum BehaviorMode : byte
